Implemented new classes
Location (in case we decide to use it),
ClueItem,
and ClueItemController.

It did mess up the scripts on a few of the clue item objects (I fixed only the lamp), though. Shouldn't be too hard to fix those, though. Just add ClueItemController scripts to them, and fill in the fields.