Clues can be found inside items
Found a good treasure chest model, put a clue in it, and after a while, managed to make it all work together.

Click the chest to zoom in on it and open it, and click the clue inside it to inspect it. Right click to stop inspecting, and then press Escape to back away from the chest.