Clues inspectable in drawer
When you click the drawer with the mirror in it, you can choose whether to inspect it or leave it alone.

If you choose to inspect the clue, everything pops up as it should. You can even add the clue to the clue log, provided the clue is in the database.

Also edited the ClueInfo struct to implement the IEquatable interface, so we can avoid duplicates in the clue log.

Also renamed all singletons named "instance" to "S".